feat(ajustes localização geográfica)

+++ DTO/LocalizacaoGeograficaDTO.cs
@@ -4,8 +4,8 @@ namespace BibliotecaViva.DTO
 {
     public class LocalizacaoGeograficaDTO : BaseDTO, IDisposable
     {
-        public long Latitude { get; set; }
-        public long Longitude { get; set; }
+        public double Latitude { get; set; }
+        public double Longitude { get; set; }
 
         public void Dispose()
         {
