+++ DTO/RegistroDTO.cs
@@ -25,8 +25,8 @@ namespace BibliotecaViva.DTO
         public string Conteudo { get; set; }
         public string Descricao { get; set; }
         public DateTime DataInsercao { get; set; }
-        public long? Latitude { get; set; }
-        public long? Longitude { get; set; } 
+        public double? Latitude { get; set; }
+        public double? Longitude { get; set; } 
         public List<ReferenciaDTO> Referencias { get; set; }
 
         public void Dispose()
